Also return ES "caused_by" error message for failed searches
If you search on a field that is mapped to a numeric,
you won't see the real reason on why this has failed.
I assume the error format has changed in recent ES versions,
and the info is now hidden in the `caused_by` field on shard failures.

For reference, this is the response:
```
{
  "took" : 2,
  "timed_out" : false,
  "_shards" : {
    "total" : 64,
    "failures" : [
      {
        "shard" : 0,
        "index" : "graylog_0",
        "reason" : {
          "type" : "query_shard_exception",
          "reason" : "failed to create query: {\n [....]  }\n}",
          "index" : "graylog_0",
          "caused_by" : {
            "type" : "number_format_exception",
            "reason" : "For input string: \"foo\""
          }
        }
      }
    ]
  },
  "hits" : {
    "total" : 0,
    "max_score" : null,
    "hits" : [ ]
  }
}
```